Move /core/src/types/state to /core/src/state
