+++ packages/core/src/reducers/core/communication/index.js
@@ -1,7 +1,7 @@
 // @flow
 import { combineReducers } from "redux-immutable";
 import { kernelspecs } from "./kernelspecs";
-import { makeCommunicationRecord } from "../../../types/state/communication";
+import { makeCommunicationRecord } from "../../../state/communication";
 
 export const communication = combineReducers(
   {
