+++ packages/core/src/state/entities/hosts.js
