+++ packages/core/src/state/entities/kernelspecs.js
