--- packages/core/src/types/index.js
@@ -1,1 +1,0 @@
-// @flow
